Fix problems

+++ pyuwsgi/server.py
@@ -32,68 +32,54 @@ class Server(object):
         self.max_requests = max_requests
         self.max_lifetime = max_lifetime
 
-        self.workers = []
+        self.workers = {}
 
-        self._is_stopping = False
-        self._is_stopping_gracefully = False
+        self.is_stopping = False
 
         self._signal_queue = []
         self._signal_pipe = []
 
-    @property
-    def is_stopping(self):
-        return self._is_stopping or self._is_stopping_gracefully
+    def find_worker_id(self, pid):
+        for n, worker in self.workers.items():
+            if worker.pid == pid:
+                return n
 
     def stop(self):
         logger.info('[master] (pid %s) stopping workers...', os.getpid())
-        self._is_stopping = True
-        for worker in self.workers:
-            worker.death = time.time() + self.mercy
-            util.kill(worker.pid, signal.SIGQUIT)
+        self.is_stopping = True
+        for worker in self.workers.values():
+            if worker.pid and not worker.death:
+                worker.death = time.time() + self.mercy
+                util.kill(worker.pid, signal.SIGQUIT)
 
     def stop_gracefully(self):
         logger.info('[master] (pid %s) gracefully stopping workers...', os.getpid())
-        self._is_stopping_gracefully = True
-        for worker in self.workers:
-            worker.death = time.time() + self.mercy
-            util.kill(worker.pid, signal.SIGTERM)
+        self.is_stopping = True
+        for worker in self.workers.values():
+            if worker.pid:
+                worker.death = time.time() + self.mercy
+                util.kill(worker.pid, signal.SIGTERM)
 
-    def find_worker(self, pid):
-        return next((w for w in self.workers if w.pid == pid), None)
-
-    def spawn(self):
-        worker = self.worker_cls(
-            self.sock,
-            self.app,
-            timeout=self.timeout,
-            connection_cls=self.connection_cls,
-            handler_cls=self.handler_cls,
-        )
-
+    def spawn(self, n):
+        worker = self.workers[n]
         pid = os.fork()
-
-        worker.pid = pid
-        worker.birth = time.time()
+        worker.reset(pid)
 
         if pid:
-            logger.info('[master] (pid %s) spawning worker (pid %s)', os.getpid(), pid)
-            return worker
-
-        return_code = errors.EXIT_CODE_STOP
+            logger.info('[master] (pid %s) spawning worker %s (pid %s)', os.getpid(), n, pid)
+            return
 
         try:
             worker.run()
-        except SystemExit:
-            pass
+        except StopIteration:
+            os._exit(errors.STOPPING)
         except errors.ApplicationError:
             logger.exception('[worker] (pid %s) cannot load application', worker.pid)
-            return_code = errors.EXIT_CODE_APPLICATION_ERROR
+            os._exit(errors.APPLICATION_ERROR)
         except:
             logger.exception('[worker] (pid %s) unhandled exception', worker.pid)
-            return_code = errors.EXIT_CODE_UNHANDLED_EXCEPTION
-
-        logger.info('[worker] (pid %s) stopped', worker.pid)
-        os._exit(return_code)
+            os._exit(errors.UNHANDLED_EXCEPTION)
+        os._exit(errors.STOPPING)
 
     def signal(self, signum, frame):
         if len(self._signal_queue) < 10:
@@ -104,20 +90,35 @@ class Server(object):
                 if e.errno not in [errno.EAGAIN, errno.EINTR]:
                     raise
 
-    def run(self):
-        logger.info('[master] (pid %s) running', os.getpid())
+    def setup_signals(self):
+        if self._signal_pipe:
+            map(os.close, self._signal_pipe)
 
         self._signal_pipe = os.pipe()
-        map(util.set_not_blocking, self._signal_pipe)
-        # map(util.set_close_on_exec, self._signal_pipe)
 
-        for signame in ['SIGINT', 'SIGQUIT', 'SIGTERM']:
+        map(util.set_not_blocking, self._signal_pipe)
+        #  map(util.set_close_on_exec, self._signal_pipe)
+
+        for signame in ['SIGINT', 'SIGQUIT', 'SIGTERM', 'SIGCHLD']:
             signum = getattr(signal, signame)
             signal.signal(signum, self.signal)
 
-        for i in range(self.procnum):
-            worker = self.spawn()
-            self.workers.append(worker)
+    def setup_workers(self):
+        for n in range(self.procnum):
+            self.workers[n] = self.worker_cls(
+                self.sock,
+                self.app,
+                timeout=self.timeout,
+                connection_cls=self.connection_cls,
+                handler_cls=self.handler_cls,
+            )
+            self.spawn(n)
+
+    def run(self):
+        logger.info('[master] (pid %s) running', os.getpid())
+
+        self.setup_signals()
+        self.setup_workers()
 
         try:
             while True:
@@ -131,6 +132,8 @@ class Server(object):
                         self.stop()
                     elif signum == signal.SIGTERM:
                         self.stop_gracefully()
+                    elif signum == signal.SIGCHLD:
+                        pass
                     else:
                         logger.warning('[master] (pid %s) ignoring signal %s', os.getpid(), signum)
 
@@ -148,51 +151,58 @@ class Server(object):
                 self.check_state()
                 self.check_children()
                 self.check_deadlines()
-        except SystemExit:
-            raise
+        except StopIteration:
+            pass
         except:
             logger.exception('[master] (pid %s) unhandled exception', os.getpid())
-            for worker in self.workers:
+            for worker in self.workers.values():
                 util.kill(worker.pid, signal.SIGKILL)
-            sys.exit(errors.EXIT_CODE_UNHANDLED_EXCEPTION)
-        finally:
-            logger.info('[master] (pid %s) stopped', os.getpid())
+            os._exit(errors.UNHANDLED_EXCEPTION)
+        os._exit(errors.STOPPING)
 
     def check_state(self):
-        if self.is_stopping and not len(self.workers):
-            raise SystemExit(errors.EXIT_CODE_STOP)
-
-        if len(self.workers) < self.procnum:
-            worker = self.spawn()
-            self.workers.append(worker)
+        if self.is_stopping:
+            for worker in self.workers.values():
+                if worker.pid:
+                    return
+            raise StopIteration
 
     def check_children(self):
         try:
             pid, status = os.waitpid(-1, os.WNOHANG)
-            worker = self.find_worker(pid)
-            if worker:
-                self.workers.remove(worker)
+
+            if pid > 0:
+                worker_id = self.find_worker_id(pid)
+                worker = self.workers.get(worker_id)
+
+                if worker:
+                    worker.pid = 0
+                    logger.info('[master] (pid %s) buried worker (pid %s)', os.getpid(), pid)
+
+                    if not self.is_stopping:
+                        self.spawn(worker_id)
+                else:
+                    logger.warning('[master] (pid %s) buried unknown process', os.getpid())
+
         except OSError as e:
             if e.errno not in [errno.EINTR, errno.ECHILD]:
                 raise
 
     def check_deadlines(self):
-        now = time.time()
-        for worker in self.workers:
-            if worker.death and worker.death < now:
-                logger.info('[worker] (pid %s) death time', worker.pid)
-                util.kill(worker.pid, signal.SIGKILL)
-                continue
+        for worker in self.workers.values():
+            if worker.pid:
+                if worker.death and worker.death < time.time():
+                    logger.info('[worker] (pid %s) death time', worker.pid)
+                    util.kill(worker.pid, signal.SIGKILL)
+                    continue
 
-            if self.max_requests and worker.requests > self.max_requests:
-                if not worker.death:
+                if not worker.death and self.max_requests and worker.requests > self.max_requests:
                     logger.info('[worker] (pid %s) requests limit', worker.pid)
                     worker.death = time.time() + self.mercy
                     util.kill(worker.pid, signal.SIGTERM)
                     continue
 
-            if self.max_lifetime and worker.lifetime > self.max_lifetime:
-                if not worker.death:
+                if not worker.death and self.max_lifetime and worker.lifetime > self.max_lifetime:
                     logger.info('[worker] (pid %s) lifetime limit', worker.pid)
                     worker.death = time.time() + self.mercy
                     util.kill(worker.pid, signal.SIGTERM)
